thinker/requirements.txt
Flask~=1.0.2
pyserial~=3.5
confluent-kafka~=2.3.0
pyserial~=3.5